requirements.txt
boto
datetime
deta
gunicorn
kaleido
numpy
pandas
plotly
python-dotenv
pytz
requests
streamlit
streamlit-authenticator==0.1.5
streamlit-option-menu
ta